handle end game event on backend (#163)

+++ backend/src/main/kotlin/com/retypeme/project/statistic/model/StatisticModel.kt
@@ -6,7 +6,7 @@ data class StatisticModel(
     val userId: String,
     val completedDuels: Int,
     val averageSpeed: Double,
-    val totalReward: Double,
+    val totalReward: Int,
     val overallWinsInDuels: Int,
     val maxSpeed: Double,
     val topSpeeds: List<Double>
